refactor(ui-rewrite): move usePromptPreview into hooks/ (#5448)
The hook and its test previously lived alongside the prompt components,
but the app convention is that reusable hooks live in `client/src/hooks/`
(useToolForm, useResourceForm, usePromptForm, useQuery, ...). Move both
files, update the four consumer imports to the `@/hooks` alias, and drop
the barrel re-exports from `components/prompts/index.ts` (no external
consumer used them).
