+++ client/src/components/prompts/index.ts
@@ -6,5 +6,3 @@ export { PromptArgsForm } from "./PromptArgsForm";
 export { PromptSnippetTabs } from "./PromptSnippetTabs";
 export { PromptPreviewButton } from "./PromptPreviewButton";
 export { PromptPreviewResult } from "./PromptPreviewResult";
-export { usePromptPreview } from "./usePromptPreview";
-export type { PromptPreviewState } from "./usePromptPreview";
