+++ client/src/hooks/usePromptPreview.test.tsx
