+++ client/src/hooks/usePromptPreview.ts
